+++ src/main/java/io/github/dataforseo/client/model/BacklinksIdListResultInfo.java
@@ -154,9 +154,9 @@ public class BacklinksIdListResultInfo  {
 
   public static final String SERIALIZED_NAME_COST = "cost";
   @SerializedName(SERIALIZED_NAME_COST)
-  private Integer cost;
+  private Double cost;
 
-  public BacklinksIdListResultInfo cost(Integer cost) {
+  public BacklinksIdListResultInfo cost(Double cost) {
     this.cost = cost;
     return this;
   }
@@ -166,11 +166,11 @@ public class BacklinksIdListResultInfo  {
    * @return cost
    */
   @javax.annotation.Nullable
-  public Integer getCost() {
+  public Double getCost() {
     return cost;
   }
 
-  public void setCost(Integer cost) {
+  public void setCost(Double cost) {
     this.cost = cost;
   }
